# library_api_gateway/api/docs/swagger.yaml
definitions:
...
  models.AllProBookList:
    properties:
      bk_authorname:
        type: string
      bk_id:
        type: string
      bk_name:
        type: string
      bk_numsbook:
        type: integer
      book_list_id:
        type: integer
      deadline:
        type: string
      given_date:
        type: string
      professor_id:
        type: string
      status:
        type: string
    type: object
...
  models.GetAllProfessorLibraryResponse:
    properties:
      count:
        type: integer
      professor_library:
        items:
          $ref: '#/definitions/models.Professor'
        type: array
    type: object
  models.GetAllProfessorResponse:
    properties:
      count:
        type: integer
      professors:
        items:
          $ref: '#/definitions/models.Professor'
        type: array
    type: object
...
  models.GetProfessorLibraryInfo:
    properties:
      professor_library:
        items:
          $ref: '#/definitions/models.Professor'
        type: array
    type: object
...
  models.ProBookList:
    properties:
      bk_id:
        type: string
      deadline:
        type: string
      professor_id:
        type: string
    type: object
  models.Professor:
    properties:
      all_booklist:
        items:
          $ref: '#/definitions/models.AllProBookList'
        type: array
      professor_firstname:
        type: string
      professor_id:
        type: string
      professor_lastname:
        type: string
      professor_phone1:
        type: string
      professor_phone2:
        type: string
    type: object
  models.ResponseModel:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/probooklist:
    get:
      consumes:
      - application/json
      description: Get Professor BookList
      operationId: get-professor-booklist
      parameters:
      - description: professor_id
        in: query
        name: professor_id
        type: string
      - description: professor_firstname
        in: query
        name: professor_firstname
        type: string
      - description: professor_lastname
        in: query
        name: professor_lastname
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.GetProfessorLibraryInfo'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get professor booklist
      tags:
      - professorbooklist
    post:
      consumes:
      - application/json
      description: Create Professor BookList
      operationId: create-professor-booklist
      parameters:
      - description: probook_list
        in: body
        name: probook_list
        required: true
        schema:
          $ref: '#/definitions/models.ProBookList'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: create professor booklist
      tags:
      - professorbooklist
    put:
      consumes:
      - application/json
      description: Return ProBook
      operationId: Return-probook
      parameters:
      - description: bk_id
        in: query
        name: bk_id
        required: true
        type: string
      - description: probook_list_id
        in: query
        name: probook_list_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Return probook
      tags:
      - professorbooklist
  /v1/probooklists:
    get:
      consumes:
      - application/json
      description: Get Professor BookLists
      operationId: get-professor-booklists
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.GetAllProfessorLibraryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get professor booklists
      tags:
      - professorbooklist
  /v1/professor:
    delete:
      consumes:
      - application/json
      description: Delete Professor
      operationId: delete-professor
      parameters:
      - description: professor_id
        in: query
        name: professor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: delete professor
      tags:
      - professor
    get:
      consumes:
      - application/json
      description: Get Professor
      operationId: get-professor
      parameters:
      - description: professor_id
        in: query
        name: professor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Professor'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get professor
      tags:
      - professor
    post:
      consumes:
      - application/json
      description: Create Professor
      operationId: create-professor
      parameters:
      - description: professor
        in: body
        name: professor
        required: true
        schema:
          $ref: '#/definitions/models.Professor'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: create professor
      tags:
      - professor
    put:
      consumes:
      - application/json
      description: Update Professor
      operationId: update-professor
      parameters:
      - description: professor
        in: body
        name: professor
        required: true
        schema:
          $ref: '#/definitions/models.Professor'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: update professor
      tags:
      - professor
  /v1/professors:
    get:
      consumes:
      - application/json
      description: Get Professors
      operationId: get-professors
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.GetAllProfessorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get professors
      tags:
      - professor
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"